+++ NEWS.md
@@ -21,6 +21,10 @@
   allows you to pick a subset of variables, and `add_layer_var` which lets 
   you add a variable indicating which layer it came from.
 
+* You can now use your inside voice for variable names with the new argument
+  `lower_vars` for `read_ipums_ddi()` and `read_ipums_micro()` family of functions
+  so that the variable names are lower case.
+
 # ipumsr 0.2.0
 * IPUMS Terra is now officially supported! Read raster, area or microdata extracts
   using functions `read_terra_raster()`, `read_terra_raster_list()`,